requirements.txt
Flask
celery
psycopg2-binary
configparser
gunicorn
retrying
requests
pytest-flask
